Let subclass of TDHttpClient call super to initialize from TDHttpClient
A subclass of TDHttpClient needs to override withHeaders method to
return an instance of the subclass. The new instance shares all private
fields excepting headers (config, httpClient, and objectMapper) with the
old instance. But because the fields are private, constructor of the
subclass can't call super with the private objects.

This change lets the subclass to overrie withHeaders method as
following:

    return new MyTDHttpClient(super.withHeaders(headers));